LayoutTests/media/track/track-change-event-expected.txt

Tests that the 'change' event is fired when a TextTrack's mode changes.


** Test a TextTrack created with addTextTrack().
RUN(track = video.addTextTrack('subtitles', 'test', 'en'))
RUN(track.mode = 'showing')
EXPECTED (video.textTracks.length == '1') OK
EVENT(change)
EXPECTED (event.target == '[object TextTrackList]') OK
EXPECTED (event instanceof window.Event == 'true') OK
EXPECTED (event.track == 'undefined') OK

END OF TEST
